// src/enum/doctor.enum.ts
export enum DoctorStatusEnum {
  ACTIVE = "active",
  INACTIVE = "inactive",
  DELETED = "deleted",
  LOCKED = "locked",
  BLOCKED = "blocked",
}

export enum ApprovalDoctorEnum {
  PENDING = "pending",
  APPROVED = "approved",
  REJECTED = "rejected",
}
export enum ApprovalDoctorStageEnum {
  PENSONAL = "personal",
  IDENTITY = "identity",
  EDUCATION = "education",
  SPECIALIST = "specialist",
  EXPERIENCE = "experience",
}